Fix the unit dependencies so they come up when tailscale-online comes up
When the tailscale-online.target goes down, the upload unit is brought down.
This works correctly.
However, when the tailscale-online.target unit comes back up, the autoupload unit is not brought back up.
The UpheldBy dependency ensures that when the tailscale-online.target is up, the autoupload unit will be started if it is not up.